requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26
opencv-python>=4.9
Pillow>=10.0
PyQt6>=6.6
matplotlib>=3.8
tqdm>=4.66
PyYAML>=6.0
roboflow>=1.1
transformers>=4.45
accelerate>=0.30
pandas>=2.2
scikit-learn>=1.4
